pyproject.toml
[project]
name = "ml-zoomcamp-2025"
version = "0.1.0"
description = "Churn Predictor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.120.0",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "uvicorn>=0.38.0",
    "torch>=2.5.0,<2.6",
    "torchvision>=0.20.0,<0.21",
    "torchaudio>=2.5.0,<2.6",
    "pandas>=2.3.3",
    "boto3>=1.42.18",
    "tf2onnx>=1.8.4",
    "pillow>=12.0.0",
    "onnxruntime>=1.23.2",
]

[dependency-groups]
dev = [
    "requests>=2.32.5",
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

#
# [tool.uv.workspace]
# members = [
#    "lead-scoring",
# ]
